move analysis statuses enum to constants file
