--- cypress/integration/enums/analysisStatuses.enum.ts
@@ -1,7 +1,0 @@
-export enum AnalysisStatuses {
-    notStarted = "Not started",
-    scheduled = "Scheduled",
-    inProgress = "In-progress",
-    completed = "Completed",
-    failed = "Failed",
-}
